Log progress during phrase generation
